refactor(sandbox): relocate stopped-container recovery helper to gateway-state
Move `startStoppedSandboxContainerForProbeRecovery` from connect.ts into
gateway-state.ts (which connect.ts already imports and which already owns
`getSandboxDockerRuntime`) and import `dockerStart` from its defining module
`adapters/docker/container` rather than the barrel. This keeps the source
architecture fan-in/fan-out budgets within limits (connect.ts fan-out and
adapters/docker barrel fan-in were both at their caps). No behavior change.

Refs #8967
